// src/components/common/form-rows/edit-plugin-form-row/edit-plugin-form-row.enums.tsx
export enum EditPluginFieldName {
    name = 'name',
    title = 'title',
    description = 'description',
    categories = 'categories',
    banner = 'banner',
    icon = 'icon',
    price = 'price',
    custom = 'custom',
    sale = 'sale',
    supported_versions = 'supported_versions',
    dependencies = 'dependencies',
    spigot_id = 'spigotId',
    github_link = 'githubLink',
    donation_url = 'donationUrl',
    body = 'body'
}